move emulator to test dir
