+++ internal/cli/cmd/agentic_tool_emulator.go
@@ -24,9 +24,9 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/g8e-ai/g8e/internal/constants"
-	clientpkg "github.com/g8e-ai/g8e/internal/agentic_tool_emulator/client"
-	"github.com/g8e-ai/g8e/internal/agentic_tool_emulator/config"
-	"github.com/g8e-ai/g8e/internal/agentic_tool_emulator/scenarios"
+	clientpkg "github.com/g8e-ai/g8e/test/agentic_tool_emulator/client"
+	"github.com/g8e-ai/g8e/test/agentic_tool_emulator/config"
+	"github.com/g8e-ai/g8e/test/agentic_tool_emulator/scenarios"
 )
 
 var (
