+++ test/agentic_tool_emulator/client/client.go
@@ -20,7 +20,7 @@ import (
 	"time"
 
 	"github.com/g8e-ai/g8e/internal/constants"
-	"github.com/g8e-ai/g8e/internal/agentic_tool_emulator/config"
+	"github.com/g8e-ai/g8e/test/agentic_tool_emulator/config"
 )
 
 // Persona is the identity Agentic Tool Emulator wears for a given exchange. This is the ONLY
