+++ test/agentic_tool_emulator/client/envelope.go
