+++ test/agentic_tool_emulator/client/protocols.go
