+++ test/agentic_tool_emulator/scenarios/governance.go
@@ -7,7 +7,7 @@ import (
 	"context"
 	"errors"
 
-	clientpkg "github.com/g8e-ai/g8e/internal/agentic_tool_emulator/client"
+	clientpkg "github.com/g8e-ai/g8e/test/agentic_tool_emulator/client"
 )
 
 // GovKit carries the mock cryptographic actors the governance scenarios need.
