+++ test/agentic_tool_emulator/scenarios/governance_test.go
@@ -6,7 +6,7 @@ package scenarios
 import (
 	"testing"
 
-	clientpkg "github.com/g8e-ai/g8e/internal/agentic_tool_emulator/client"
+	clientpkg "github.com/g8e-ai/g8e/test/agentic_tool_emulator/client"
 )
 
 func TestSetGovKit(t *testing.T) {
