+++ test/agentic_tool_emulator/scenarios/scenario.go
@@ -12,7 +12,7 @@ import (
 	"os"
 	"time"
 
-	clientpkg "github.com/g8e-ai/g8e/internal/agentic_tool_emulator/client"
+	clientpkg "github.com/g8e-ai/g8e/test/agentic_tool_emulator/client"
 )
 
 // Posture is the Gateway enforcement mode a scenario needs.
