+++ test/agentic_tool_emulator/scenarios/scenario_test.go
@@ -11,7 +11,7 @@ import (
 	"testing"
 	"time"
 
-	clientpkg "github.com/g8e-ai/g8e/internal/agentic_tool_emulator/client"
+	clientpkg "github.com/g8e-ai/g8e/test/agentic_tool_emulator/client"
 )
 
 func TestRegistry(t *testing.T) {
